Remove EmailFields concern
