+++ app/mailers/notification_mailer.rb
@@ -60,4 +60,20 @@ private
       "List-Id" => list_id(post.thread.subforum)
     )
   end
+
+  def reply_to_post_address(reply_info)
+    "Community <reply-#{reply_info}@mail.community.hackerschool.com>"
+  end
+
+  def list_address(subforum)
+    "#{subforum.name.downcase.gsub(/\s+/, '-')}@lists.community.hackerschool.com"
+  end
+
+  def subforum_thread_subject(thread)
+    "[Community - #{thread.subforum.name}] #{thread.title}"
+  end
+
+  def list_id(subforum)
+    "<#{subforum.name.downcase.gsub(/\s+/, '-')}.community.hackerschool.com>"
+  end
 end
